+++ src/components/EventViewer/AnnotationUI/Annotations/Annotation.astro
@@ -1,7 +1,7 @@
 ---
 import RichText from '@components/RichText/index.astro';
 import { PlayCircleIcon } from '@heroicons/react/24/outline';
-import TagPill from '@components/Tags/TagPill';
+import TagPill from '@components/TagUtils/TagPill';
 import { formatTimestamp } from 'src/utils/player';
 import { getEntry } from 'astro:content';
 import type { Annotation } from '@ty/index';
