+++ src/components/EventViewer/AnnotationUI/TagFilter.tsx
@@ -1,4 +1,4 @@
-import TagPill from '@components/Tags/TagPill.tsx';
+import TagPill from '@components/TagUtils/TagPill.tsx';
 import {
   Checkbox,
   Popover,
