+++ src/components/TagUtils/AnnotationBand.astro
@@ -1,6 +1,6 @@
 ---
 import type { Annotation, TagGroup } from '@ty/index.ts';
-import TagPill from './TagPill.tsx';
+import TagPill from './TagPill.js';
 import { formatTimestamp } from 'src/utils/player.ts';
 import RichText from '../RichText/index.astro';
 import { getCategoryLabel } from 'src/utils/string';
