+++ src/pages/tags/detail/[category].astro
@@ -1,8 +1,8 @@
 ---
 import Breadcrumbs from '@components/Breadcrumbs.astro';
 import Container from '@components/Container.astro';
-import AnnotationSpectrum from '@components/Tags/AnnotationSpectrum.astro';
-import TagPill from '@components/Tags/TagPill.tsx';
+import AnnotationSpectrum from '@components/TagUtils/AnnotationSpectrum.astro';
+import TagPill from '@components/TagUtils/TagPill.tsx';
 import { CheckIcon } from '@heroicons/react/24/outline';
 import Layout from '@layouts/Layout.astro';
 import type { GetStaticPaths } from 'astro';
