+++ src/pages/tags/detail/index.astro
@@ -5,8 +5,8 @@ import { dynamicConfig } from '../../../../dynamic-astro-config';
 import Breadcrumbs from '@components/Breadcrumbs.astro';
 import { getEntry } from 'astro:content';
 import { getCollection } from 'astro:content';
-import AnnotationSpectrum from '@components/Tags/AnnotationSpectrum.astro';
-import TagPill from '@components/Tags/TagPill';
+import AnnotationSpectrum from '@components/TagUtils/AnnotationSpectrum.astro';
+import TagPill from '@components/TagUtils/TagPill';
 import { CheckIcon } from '@heroicons/react/24/outline';
 import { getCategoryLabel } from 'src/utils/string';
